Moves TypeFilter into helper
