+++ src/app/modules/item/components/composition-filter/composition-filter.component.ts
@@ -1,7 +1,6 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { ensureDefined } from 'src/app/shared/helpers/assert';
-
-export type TypeFilter = 'Groups'|'Teams'|'Users';
+import { TypeFilter } from '../../helpers/composition-filter';
 
 @Component({
   selector: 'alg-composition-filter',
