+++ src/app/modules/item/helpers/permissions-constraints-validator.ts
@@ -1,7 +1,7 @@
 import { ValidationErrors, ValidatorFn, AbstractControl } from '@angular/forms';
 import { GroupPermissions } from 'src/app/shared/http-services/group-permissions.service';
 import { ItemCorePerm } from 'src/app/shared/models/domain/item-permissions';
-import { TypeFilter } from '../components/composition-filter/composition-filter.component';
+import { TypeFilter } from './composition-filter';
 import {
   validateCanView,
   validateCanGrantView,
