+++ src/app/modules/item/helpers/permissions-texts.ts
@@ -3,7 +3,7 @@ import {
 } from 'src/app/modules/shared-components/components/collapsible-section/progress-select/progress-select.component';
 import { GroupPermissions } from 'src/app/shared/http-services/group-permissions.service';
 import { generateErrorMessage, permissionsInfoString } from './permissions-string';
-import { TypeFilter } from '../components/composition-filter/composition-filter.component';
+import { TypeFilter } from './composition-filter';
 import {
   validateCanView,
   validateCanGrantView,
